% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_size.R
\name{effect_cont_emp}
\alias{effect_cont_emp}
\title{Association Size of Continuous Variable T}
\usage{
effect_cont_emp(params, dat5)
}
\arguments{
\item{params}{a numeric vector containing the estimated correlation parameters in the CoCoA model}

\item{dat5}{a data frame containing only the covariates in the correlation model (i.e., excluding X and Y)}
}
\value{
A numeric data frame containing the marginal and conditional association sizes for each
unique value of T, as well as the estimated parameters.
}
\description{
Association Size of Continuous Variable T
}
\details{
The data frame \code{dat5} may contain a column with name "1" consisting of 1's
corresponding to the intercept term in the correlation model. If not, this column
will be created. Currently, only the \code{tanh} link is implemented.
}
\examples{
T = sample(x = seq(from = 0, to = 2, by = 0.25), size = 100, replace = TRUE)
Z = rep(NA, times = 100)
for (i in 1:100){
Z[i] <- if (T[i] < 1){
# Z|T=1 ~ Bernoulli(0.3)
as.numeric(purrr::rbernoulli(n = 1, p = 0.3))
} else {
# Z|T=0 ~ Bernoulli(0.9)
as.numeric(purrr::rbernoulli(n = 1, p = 0.9))
}
}
dat_zt0 <- data.frame(Z = Z, T = T)
effect_cont_emp(params = c(0.5, 0.2, 0.5), dat5 = dat_zt0)

}
