% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_reml.R
\name{format_predictors}
\alias{format_predictors}
\title{Format data into longer format}
\usage{
format_predictors(dat_xyz)
}
\arguments{
\item{dat_xyz}{a data frame with columns containing observations of X, Y, ... (in that order)}
}
\description{
Format data into longer format
}
\details{
Given a data frame \code{dat_xyz}, reformat it so that there is 1 row for each X and 1 row for each Y observation per person.
Note: this function is an adapted version of the form() function from the LiquidAssociation package; see
\url{https://www.bioconductor.org/packages/release/bioc/html/LiquidAssociation.html}.
}
