% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_cop.R
\name{get_beta_params}
\alias{get_beta_params}
\title{Beta maximum likelihood estimator}
\usage{
get_beta_params(dat_x)
}
\arguments{
\item{dat_x}{a numeric vector of observations}
}
\value{
The maximum likelihood estimates of the Beta parameters given the data \code{dat_x}.
}
\description{
Beta maximum likelihood estimator
}
\keyword{internal}
