% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_cop.R
\name{get_gamma_params}
\alias{get_gamma_params}
\title{Gamma maximum likelihood estimator}
\usage{
get_gamma_params(dat_x)
}
\arguments{
\item{dat_x}{a numeric vector of observations}
}
\value{
The maximum likelihood estimates of the Gamma parameters given the data \code{dat_x}.
}
\description{
Gamma maximum likelihood estimator
}
\keyword{internal}
