% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_funs.R
\name{get_link_name}
\alias{get_link_name}
\title{Link function label}
\usage{
get_link_name(str)
}
\arguments{
\item{str}{character of the link function}
}
\description{
For plots/tables, get the name of the link function.
}
\examples{
ptab(cars)

}
