% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_reml.R
\name{get_params_reml}
\alias{get_params_reml}
\title{Restricted maximum likelihood (REML) estimation of CoCoA model parameters}
\usage{
get_params_reml(
  dat_xyz,
  init_mle = TRUE,
  rescale_zt = FALSE,
  optim_type = "optim"
)
}
\arguments{
\item{dat_xyz}{a data frame containing the columns X, Y, Z, T, ... in that order}

\item{init_mle}{a logical value indicating if the parameter estimates should be initialized at the MLE}

\item{rescale_zt}{a logical value indicating if the conditional correlation model covariates Z, T, ... should be rescaled (z-scored)}

\item{optim_type}{a character value. If \code{optim_type = "optim"}, the usual \code{optim} optimization routine is used.
Otherwise the "Rvmmin" method from the optimr package is used.}
}
\value{
A list containing \code{params}, a numeric matrix of the estimated parameters, \code{out2}, a table of the estimated parameters,
\code{runtime}, the runtime of the optimization routine, and \code{loglik}, the value of the final restricted log-likelihood.
}
\description{
Estimate CoCoA model parameters using the restricted maximum likelihood (REML) estimator and the multivariate
normal likelihood.
}
\details{
Given \eqn{n} observations of Z and T, X and Y are bivariate normal with
means \code{mu_x, mu_y}, standard deviations \code{sd_x, sd_y}, and correlation
\deqn{Corr(X,Y) = g^{-1}(\alpha + \beta Z + \gamma T + ...).} The inverse link function \eqn{g^{-1}}
allows for unconstrained parameter estimation while ensuring that the predicted correlation is bounded.
Currently, only the \code{tanh} link is implemented.
}
\examples{
T = as.numeric(purrr::rbernoulli(n = 100, p = 0.5))
Z = rep(NA, times = 100)
for (i in 1:100){
Z[i] <- if (T[i] == 1){
# Z|T=1 ~ Bernoulli(0.3)
as.numeric(purrr::rbernoulli(n = 1, p = 0.3))
} else {
# Z|T=0 ~ Bernoulli(0.9)
as.numeric(purrr::rbernoulli(n = 1, p = 0.9))
}
}
dat_zt0 <- data.frame(Z = Z, T = T, ZT = Z*T)
XY = sim_xy2_mvn(theta0 = c(0, 0, 0, 0), dat_zt0 = dat_zt0,
mu_x = 1, mu_y = 1, sd_x = 1, sd_y = 2, cor_link = "tanh")
dat_xyz = data.frame(X = XY$x, Y = XY$y)
dat_xyz = cbind(dat_xyz, dat_zt0)
get_params_reml(dat_xyz)

}
