% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_funs.R
\name{link_funs}
\alias{link_funs}
\alias{linkinv_tanh}
\alias{linkfun_tanh}
\alias{linkinv_identity}
\alias{linkfun_identity}
\alias{linkfun_tid}
\alias{linkinv_tid}
\alias{linkfun_logit}
\alias{linkinv_logit}
\alias{linkfun_probit}
\alias{linkinv_probit}
\alias{linkfun_cloglog}
\alias{linkinv_cloglog}
\title{Link and inverse link functions for the correlation model.}
\usage{
linkinv_tanh(eta)

linkfun_tanh(p)

linkinv_identity(eta)

linkfun_identity(p)

linkfun_tid(p)

linkinv_tid(eta)

linkfun_logit(p)

linkinv_logit(eta)

linkfun_probit(p)

linkinv_probit(eta)

linkfun_cloglog(p)

linkinv_cloglog(eta)
}
\arguments{
\item{eta}{a numeric value corresponding to the linear predictor, generally taking values in (-Inf, Inf).}

\item{p}{a numeric value corresponding to the outcome (correlation), taking values in (0,1) for \code{logit}, \code{probit}; or (-1,1) for \code{tanh}.}
}
\description{
Link and inverse link functions for the correlation model.
}
\details{
Link functions are defined according to the C++ source code in geepack \url{https://github.com/cran/geepack/blob/master/src/famstr.cc}.
Both \code{linkfun_tid()} and \code{linkinv_tid()} correspond to the truncated identity link.
}
\examples{
tanh(3)
atanh(0.5)
}
