% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_cop.R
\name{loglik_beta}
\alias{loglik_beta}
\title{Beta log-likelihood}
\usage{
loglik_beta(params = c(1, 1), dat_x)
}
\arguments{
\item{params}{a numeric vector of length 2 for the Beta shape parameters}

\item{dat_x}{a numeric vector of observations}
}
\value{
The Beta log-likelihood given the data \code{dat_x} and parameters \code{params}.
}
\description{
Beta log-likelihood
}
