% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_cop.R
\name{loglik_gamma}
\alias{loglik_gamma}
\title{Gamma log-likelihood}
\usage{
loglik_gamma(params = c(1, 1), dat_x)
}
\arguments{
\item{params}{a numeric vector of length 2 for the Gamma shape parameters}

\item{dat_x}{a numeric vector of observations}
}
\value{
The Gamma log-likelihood given the data \code{dat_x} and parameters \code{params}.
}
\description{
Gamma log-likelihood
}
