% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_funs.R
\name{nl}
\alias{nl}
\title{Print new line in HTML files}
\usage{
nl()
}
\description{
This function prints a new line in HTML files, which is useful when printing other things such as headers or plots.
}
\examples{
nl()

}
