% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_funs.R
\name{ptab}
\alias{ptab}
\title{Output data.frame as pandoc.table()}
\usage{
ptab(x)
}
\arguments{
\item{x}{a data.frame}
}
\description{
This is a wrapper around pandoc.table() that ensures all output will be printed on the same line.
}
\examples{
ptab(cars)

}
\keyword{internal}
