% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_reml.R
\name{reml_get_beta}
\alias{reml_get_beta}
\title{Estimate mean parameters from the REML likelihood}
\usage{
reml_get_beta(pars, dat_xyz)
}
\arguments{
\item{pars}{a numeric vector containing the standard deviation of X and Y, followed by the conditional correlation parameters}

\item{dat_xyz}{a data frame containing observations of X, Y, Z, ... in that order}
}
\value{
A numeric vector of the REML estimates of the mean parameters.
}
\description{
Estimate mean parameters from the REML likelihood
}
\details{
Given an estimate of the variance and correlation parameters, estimate the mean parameters using
restricted maximum likelihood.
}
\keyword{internal}
