% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_funs.R
\name{round2}
\alias{round2}
\title{Round and format numbers for table output}
\usage{
round2(x, digits)
}
\arguments{
\item{x}{a numeric value}

\item{digits}{an integer indicating the number of digits to display}
}
\value{
The formatted number (as a character).
}
\description{
Converts a numeric value to a character value with the specified number of digits after the decimal.
}
\examples{
round(0.100, digits = 2)
round2(0.100, digits = 2)

}
\keyword{internal}
