% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_funs.R
\name{scale2}
\alias{scale2}
\title{Scale vector using mean and standard deviation}
\usage{
scale2(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical value to be passed to mean() and sd()}
}
\value{
The scaled vector.
}
\description{
Linearly rescale a vector to have mean = 0 and sd = 1.
}
\examples{
scale2(1:10, na.rm = TRUE)

}
\keyword{internal}
