% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sim_xy2_cop}
\alias{sim_xy2_cop}
\title{Simulate copula-valued data in CoCoA model}
\usage{
sim_xy2_cop(theta0, dat_zt0, cor_link = "tanh")
}
\arguments{
\item{theta0}{numeric vector of true parameters}

\item{dat_zt0}{data.frame containing observations of Z and T}

\item{cor_link}{character vector corresponding to the correlation link function (default = "tanh")}
}
\value{
Generated X and Y values based on the copula model.
}
\description{
Simulate (X,Y) from a Gaussian copula with Beta and Gamma margins given observations of (Z,T) and the CoCoA model.
}
\details{
Given \eqn{n} observations of Z and T, generate X and Y according to the model \deqn{Corr_s(X,Y) = g^{-1}(\alpha_s + \beta_s Z + \gamma_s T + ...)}
where \eqn{Corr_s(X,Y)} is the Spearman correlation, X follows a Gamma(2,1) marginal, and Y follows a Beta(2,2) marginal.
}
\examples{
T = as.numeric(purrr::rbernoulli(n = 100, p = 0.5))
Z = rep(NA, times = 100)
for (i in 1:100){
Z[i] <- if (T[i] == 1){
# Z|T=1 ~ Bernoulli(0.3)
as.numeric(purrr::rbernoulli(n = 1, p = 0.3))
} else {
# Z|T=0 ~ Bernoulli(0.9)
as.numeric(purrr::rbernoulli(n = 1, p = 0.9))
}
}
dat_zt0 <- data.frame(Z = Z, T = T, ZT = Z*T)
XY = sim_xy2_cop(theta0 = c(0, 0, 0, 0), dat_zt0 = dat_zt0, cor_link = "tanh")


}
