% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{true_params_cop}
\alias{true_params_cop}
\title{Obtain Pearson correlation parameters from copula-based CoCoA model (internal function)}
\usage{
true_params_cop(
  alpha = 0.5,
  beta = 0.8,
  gamma = 0.1,
  delta = 0.5,
  cor_link = "tanh"
)
}
\arguments{
\item{alpha}{numeric parameter in copula model (see details)}

\item{beta}{numeric parameter in copula model (see details)}

\item{gamma}{numeric parameter in copula model (see details)}

\item{delta}{numeric parameter in copula model (see details)}

\item{cor_link}{character vector corresponding to the correlation link function (default = "tanh")}
}
\value{
A numeric matrix containing the model parameters corresponding to the Pearson correlation model.
}
\description{
Obtain Pearson correlation parameters from copula-based CoCoA model (internal function)
}
\details{
Given \eqn{n} observations of binary Z and T, and X and Y generated according to the model \deqn{Corr_s(X,Y) = g^{-1}(\alpha_s + \beta_s Z + \gamma_s T + \delta_s Z*T)}
where \eqn{Corr_s(X,Y)} is the Spearman correlation, X follows a Gamma(2,1) marginal, and Y follows a Beta(2,2) marginal, recover the parameters
\eqn{(\alpha, \beta, \gamma, \delta)} corresponding to the *Pearson* correlation model \deqn{Corr(X,Y) = g^{-1}(\alpha + \beta Z + \gamma T + \delta Z*T).}
}
\examples{
theta_true = true_params_cop(0, 0, 0, 0, cor_link = "tanh")

}
