% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_cop.R
\name{bound_away}
\alias{bound_away}
\title{Bound values away from overly large and small values}
\usage{
bound_away(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
The bounded value.
}
\description{
Bound values away from overly large and small values
}
\details{
Bounds the value \code{x} away from
}
